refactor: Remove PDI
