+++ app/api/endpoints/person.py
@@ -5,8 +5,8 @@ from pymongo.collection import Collection
 from fastapi import APIRouter, Depends, HTTPException
 
 from app.db.database import get_db
-from app.models.schemas import PersonResponse, PersonBase, PersonUpdate
-from app.services.person_service import create_person, get_person, get_all_person, update_person, delete_person
+from app.models.schemas.person import PersonResponse, PersonBase, PersonUpdate
+from app.services.person import create_person, get_person, get_all_person, update_person, delete_person
 
 router = APIRouter()
